installed active scaffold version 1.2RC1
installed the release candidate 1 of active scaffold 1.2 using git
here is a copy of the console session:

> git clone git://github.com/activescaffold/active_scaffold.git vendor/plugins/active_scaffold
> cd vendor/plugins/active_scaffold
> git checkout 1.2RC1
> rm -rf .git
> cd ../../..
> script/runner vendor/plugins/active_scaffold/install.rb

Active Scaffold is a Rails plugin (under MIT license) that makes easy
to create a good paginated AJAX-based CRUD interface which can be customized
overriding the default parameters and views.

Homepage: http://activescaffold.com/

The idea is to use the interface for file management, both public files and own files.