added a column to "users" to cache the public annotations count
The column "public_annotations_count" was added to cache the number of public
annotations present in each record. This simplifies and speeds up the process
of selecting those users which have a public annotation for listing in the
public repository. On the other side it must be incremented/decremented each
time the status of an annotation is changed (see modification of file_manager.rb).

If this is the only application accessing the database, the column will
contain meaningful data. If other applications write in the database, they
must respect this cache column.