added failsafe against gt server crash
SequenceRegion#to_png will now continue forever to try to get an image from
the gt server.

If the gt server is called in an infinite loop, restarting itself after a
failure, the whole thing will be almost transparent (apart from the delay)
for the user.

script/loop script/gt_server

implements it for example with a minimal bash script.