Viewer settings: width text field now working;
default value was set to 900 for users that are not logged in
or to the own default settings for logged in users

The width may be specified in the URL (i.e. ?width=900)
but it's sent by the settings form as a
POST parameter (i.e. not encoded in the URL).