(MIGRATION): created table to log UUIDs
When a page containing an image (Viewer#index) is loaded,
an UUID is created. The image is generated on the GTServer
using the necessary parameters and the UUID as identifier.
The image is then loaded using as URL "image/#{uuid}".
As the GTServer is restarted each time the memory is full,
there is a concrete possibility that between the time the
image is generated and requested, the server is restarted.
If this happens then the #image method will use the log
to reconstruct the parameters and repeat the generation
request before the image is sent to the browser.