remove major performance bottleneck in image map generation
The major problem was that C objects cannot be marshalled by DRb (obviously).
Therefore, when image_info.each_hotspot was called each request would be
handed over to the gt_server which returned the coordinates of a hotspot and a
proxy for the corresponding feature. Each call on the feature (like
feature.get_score) would produce another DRb call. All this DRb calls where
quite time consuming.

Solution: The ImageInfo class of gtruby was extended with a get_hotspots()
methods which collects all the necessary information at once.
With this method a single DRb call sends everything at once (it can be
marshalled now).
Because this method is somewhat special and is probably only useful in the
context of GenomeViewer, I put it in the gt_server himself (Ruby's open
classes made this possible).

Some ideas for further performance improvements:
- use a singe array/hash for a hotspot instead of an array with an embedded hash
- reduce the amount of displayed information
- inline some methods which build the image map
- use arrays to build string before concatenating them:

  s = []
  s << "multiple times"
  s.join

  instead of

  s = ""
  s << "multiple times"
  s

But only profiling can tell if this ideas work and are worth the effort.

[#2 state:resolved]